requirements.txt
coverage==7.6.8
paho-mqtt==2.1.0
